+++ src/ingest/tests.rs
@@ -44,7 +44,7 @@ const KEY_PATH: &str = "tests/key.pem";
 const CA_CERT_PATH: &str = "tests/root.pem";
 const HOST: &str = "localhost";
 const TEST_PORT: u16 = 60190;
-const PROTOCOL_VERSION: &str = "0.12.3";
+const PROTOCOL_VERSION: &str = "0.13.0";
 
 struct TestClient {
     conn: Connection,
@@ -823,13 +823,13 @@ async fn tls() {
         cipher: 10,
         ja3s: "ja3s".to_string(),
         serial: "serial".to_string(),
-        subject_country: "sub_contry".to_string(),
+        subject_country: "sub_country".to_string(),
         subject_org_name: "sub_org".to_string(),
         subject_common_name: "sub_comm".to_string(),
         validity_not_before: 11,
         validity_not_after: 12,
         subject_alt_name: "sub_alt".to_string(),
-        issuer_country: "issuer_contry".to_string(),
+        issuer_country: "issuer_country".to_string(),
         issuer_org_name: "issuer_org".to_string(),
         issuer_org_unit_name: "issuer_org_unit".to_string(),
         issuer_common_name: "issuer_comm".to_string(),
@@ -868,7 +868,7 @@ async fn smb() {
         last_time: 1,
         command: 0,
         path: "something/path".to_string(),
-        service: "sevice".to_string(),
+        service: "service".to_string(),
         file_name: "fine_name".to_string(),
         file_size: 10,
         resource_type: 20,